refactoring existed tasks
