+++ .gulp/deploy.js
@@ -1,0 +1,35 @@
+/**
+ * Deploying app to server
+ */
+
+var gulp = require('gulp'),
+    minimist = require('minimist'),
+    ftp = require('vinyl-ftp'),
+    gutil = require('gulp-util'),
+    debug = require('gulp-debug');
+
+var args = minimist(process.argv.slice(2));
+
+gulp.task('deploy', function () {
+    var isPull = args.request !== 'false';
+    console.log(args.request);
+    console.log('Is pull-request: ' + isPull);
+    console.log('Triggering deployment: ' + !isPull);
+    if (!isPull) {
+        console.log('Deploying...');
+        var remotePath = './';
+        var conn = ftp.create({
+            host: 'ftp.s5.domain-ellenorzes.hu',
+            user: args.user,
+            password: args.password,
+            log: gutil.log
+        });
+
+        gulp.src(['./src/*', './src/**/*'])
+            .pipe(debug())
+            .pipe(conn.newer(remotePath))
+            .pipe(conn.dest(remotePath));
+    } else {
+        console.log ('Deployment skipped.');
+    }
+});
