+++ package.json
@@ -4,7 +4,7 @@
   "description": "Time-tracking tool for counting working hours daily.",
   "main": "index.js",
   "scripts": {
-    "test": "node node_modules/gulp/bin/gulp.js --no-color test"
+    "test": "./node_modules/.bin/karma start --single-run --browsers PhantomJS"
   },
   "repository": {
     "type": "git",
@@ -28,6 +28,7 @@
     "gulp": "^3.9.1",
     "gulp-concat": "^2.6.0",
     "gulp-debug": "^2.1.2",
+    "gulp-jshint": "^2.0.0",
     "gulp-rename": "^1.2.2",
     "gulp-sourcemaps": "^1.6.0",
     "gulp-uglify": "^1.5.3",
@@ -40,6 +41,7 @@
     "karma-phantomjs-launcher": "^1.0.0",
     "minimist": "^1.2.0",
     "phantomjs-prebuilt": "^2.1.7",
+    "require-dir": "^0.3.0",
     "vinyl-ftp": "^0.4.5"
   }
 }
